Move aws default credentials test
