+++ tests/common/configuration/test_credentials.py
@@ -288,18 +288,15 @@ def test_run_configuration_slack_credentials(environment: Any) -> None:
     assert c.slack_incoming_hook == "DBgAXQFPQVsAAEteXlFRWUoPG0BdHQ-EbAg=="
 
 
-def test_aws_credentials_resolved_from_default(environment: Dict[str, str]) -> None:
-    environment['AWS_ACCESS_KEY_ID'] = 'fake_access_key'
-    environment['AWS_SECRET_ACCESS_KEY'] = 'fake_secret_key'
-    environment['AWS_SESSION_TOKEN'] = 'fake_session_token'
+def test_aws_credentials_resolved(environment: Dict[str, str]) -> None:
+    environment['CREDENTIALS__AWS_ACCESS_KEY_ID'] = 'fake_access_key'
+    environment['CREDENTIALS__AWS_SECRET_ACCESS_KEY'] = 'fake_secret_key'
+    environment['CREDENTIALS__AWS_SESSION_TOKEN'] = 'fake_session_token'
+    environment['CREDENTIALS__AWS_PROFILE'] = 'fake_profile'
 
     config = resolve_configuration(AwsCredentials())
 
     assert config.aws_access_key_id == 'fake_access_key'
     assert config.aws_secret_access_key == 'fake_secret_key'
     assert config.aws_session_token == 'fake_session_token'
-    assert config.aws_profile == 'default'
-
-
-if __name__ == '__main__':
-    pytest.main(['tests/common/configuration/test_credentials.py', '-k', 'aws', '--pdb'])
+    assert config.aws_profile == 'fake_profile'
